+++ lib/ansi.js
@@ -14,13 +14,14 @@ export default class AnsiStyle extends EtchComponent {
 
   render () {
     const text = this.props.text || ''
+    const map = (this.props.mapText) || ((text) => <span>{text}</span>)
 
     return (
       <div className='go-plus-panel-pre'>
         {
           parser(text).map((chunk, i) => {
             const classes = chunk.styles.map(style => 'go-plus-ansi-' + style).join(' ')
-            return <span className={classes} key={i}>{chunk.text}</span>
+            return <span className={classes} key={i}>{map(chunk.text)}</span>
           })
         }
       </div>
