+++ lib/test/test-panel.js
@@ -18,7 +18,36 @@ export default class TestPanel extends EtchComponent {
     if (this.props.model) {
       this.props.model.view = this
     }
-    this.locationRegex = /([\w]*_test.go:\d+)/g
+    
+    this.locationRegex = /([\w-/.]*.go:\d+)/g
+  }
+
+  makeLink (text) {
+    const elements = []
+    let lastIndex = 0
+
+    let match
+    if (this.locationRegex) {
+      do {
+        match = this.locationRegex.exec(text)
+        if (match && match.index) {
+          // take raw text up to this match
+          elements.push(<span>{text.slice(lastIndex, match.index)}</span>)
+
+          const linkText = match[0]
+          // convert the match to a link
+          elements.push(<a onclick={() => this.linkClicked(linkText)}>{linkText}</a>)
+          lastIndex = match.index + match[0].length
+        }
+      } while (match)
+    }
+
+    // raw text from last match to the end
+    if (lastIndex < text.length) {
+      elements.push(<span>{text.slice(lastIndex)}</span>)
+    }
+
+    return elements
   }
 
   render () {
@@ -33,7 +62,7 @@ export default class TestPanel extends EtchComponent {
 
     return (
       <div ref='content' className='go-plus-test-panel' scrollTop={this.scrollHeight} style={style}>
-        <AnsiStyle text={output} />
+        <AnsiStyle text={output} mapText={this.makeLink.bind(this)} />
       </div>
     )
   }
